requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
astroid==2.15.8
autopep8==2.0.4
bandit==1.7.5
beautifulsoup4==4.12.2
boto3==1.28.80
botocore==1.31.80
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dill==0.3.7
dj-database-url==2.1.0
Django==3.2.23
django-allauth==0.58.2
django-allauth-ui==0.1.0
django-appconf==1.0.5
django-badgify==0.6.0
django-bmemcached==0.3.0
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.1
django-select2==7.11.1
django-sslserver==0.22
django-storages==1.14.2
django-tables2==2.6.0
django-widget-tweaks==1.4.12
docutils==0.20.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
isort==5.11.5
jmespath==1.0.1
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pbr==6.0.0
Pillow==9.5.0
platformdirs==3.11.0
psycopg2==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.13.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.7
pylint-plugin-utils==0.8.2
python-binary-memcached==0.31.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
q==2.7
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
s3transfer==0.7.0
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
stevedore==3.5.2
tomli==2.0.1
tomlkit==0.12.2
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.7.1
uhashring==2.3
urllib3==2.0.7
webexteamssdk==1.6.1
whitenoise==6.5.0
wrapt==1.15.0
yapf==0.40.2
zipp==3.15.0
